Migrate first wave of models into HAModels package
Moves nine model types into the HAModels package following the pattern
established by WatchComplicationConfig: the pure struct (Foundation +
GRDB only) lives in HAModels, while Current/L10n/HAKit-dependent
helpers stay behind as extensions in their original files.

Migrated: AllowedTag, KioskSettings (+ kiosk enums), AssistConfiguration,
NotificationSnoozeAction, AppArea, AppDeviceRegistry, AssistPipelines +
Pipeline, HAAppEntity + ServerEntity, AppPanel.

Pipeline keeps its HADataDecodable conformance via a retroactive
extension in Shared. Also removes Sources/Shared/Environment/
DeviceRegistry.swift, an orphaned duplicate of DeviceRegistryEntry.swift
with no project references.

Still blocked on the MagicItem split: WatchConfig, CarPlayConfig,
MacToolbarConfig, AppIconShortcutConfig, CustomWidget.
