+++ Sources/App/Assist/AssistConfiguration.swift
@@ -2,44 +2,9 @@ import Foundation
 import GRDB
 import Shared
 
-/// Configuration for the Assist feature, persisted in the database
-struct AssistConfiguration: Codable, Identifiable, Equatable, PersistableRecord, FetchableRecord {
-    /// Singleton ID for the configuration (only one row in the database)
-    static let singletonID = "assist_config"
-
-    var id: String = AssistConfiguration.singletonID
-    var enableOnDeviceSTT: Bool = false
-    var onDeviceSTTLocaleIdentifier: String? = nil
-    var muteTTS: Bool = false
-    var enableOnDeviceTTS: Bool = false
-    var onDeviceTTSVoiceIdentifier: String? = nil
-
-    /// Custom row initializer to handle NULL values from migrated columns.
-    init(row: Row) throws {
-        self.id = row[DatabaseTables.AssistConfiguration.id.rawValue]
-        self.enableOnDeviceSTT = row[DatabaseTables.AssistConfiguration.enableOnDeviceSTT.rawValue] ?? false
-        self.onDeviceSTTLocaleIdentifier = row[DatabaseTables.AssistConfiguration.onDeviceSTTLocaleIdentifier.rawValue]
-        self.muteTTS = row[DatabaseTables.AssistConfiguration.muteTTS.rawValue] ?? false
-        self.enableOnDeviceTTS = row[DatabaseTables.AssistConfiguration.enableOnDeviceTTS.rawValue] ?? false
-        self.onDeviceTTSVoiceIdentifier = row[DatabaseTables.AssistConfiguration.onDeviceTTSVoiceIdentifier.rawValue]
-    }
-
-    init(
-        id: String = AssistConfiguration.singletonID,
-        enableOnDeviceSTT: Bool = false,
-        onDeviceSTTLocaleIdentifier: String? = nil,
-        muteTTS: Bool = false,
-        enableOnDeviceTTS: Bool = false,
-        onDeviceTTSVoiceIdentifier: String? = nil
-    ) {
-        self.id = id
-        self.enableOnDeviceSTT = enableOnDeviceSTT
-        self.onDeviceSTTLocaleIdentifier = onDeviceSTTLocaleIdentifier
-        self.muteTTS = muteTTS
-        self.enableOnDeviceTTS = enableOnDeviceTTS
-        self.onDeviceTTSVoiceIdentifier = onDeviceTTSVoiceIdentifier
-    }
-
+// `AssistConfiguration` itself lives in the `HAModels` package; these are its database-backed
+// accessors.
+extension AssistConfiguration {
     static var config: AssistConfiguration {
         do {
             return try Current.database().read { db in
