+++ Sources/Shared/HAAppEntity.swift
@@ -2,68 +2,18 @@ import Foundation
 import GRDB
 import HAKit
 
-/// The entity "universe" cache, sourced from the REST `/states` endpoint (see `AppEntitiesModel`).
-///
-/// This is distinct from, and complementary to, `EntityRegistryListForDisplay.Entity` (the registry,
-/// from `config/entity_registry/list_for_display`):
-/// - `HAAppEntity` is every entity that currently has a state — including ones with no registry entry
-///   (YAML/template/command-line entities, etc.) — and carries `domain` + `rawDeviceClass`, which the
-///   registry does not. It's what pickers/widgets enumerate as "all selectable entities".
-/// - The registry is config metadata (area, hidden, decimal precision, the user's name) for the
-///   registered, non-disabled subset, and is only consulted to filter/enrich those entities.
-///
-/// `name` holds the **resolved display name**: the registry name (`list_for_display` `en`) when the
-/// entity has a registry row, otherwise the live `friendly_name`, otherwise the `entityId`. It is
-/// resolved once, at write time, by `AppEntitiesModel` (see `handle(appRelatedEntities:server:)`), so
-/// readers can use `name` directly — there is no per-read registry lookup.
-public struct HAAppEntity: Codable, Identifiable, FetchableRecord, PersistableRecord, Equatable {
-    public let id: String
-    public let entityId: String
-    public let serverId: String
-    public let domain: String
-    /// The entity's resolved **display name**, persisted in the database. `AppEntitiesModel` populates
-    /// this at write time with the registry name (`list_for_display` `en`) when one exists, falling back
-    /// to the live `friendly_name`, then the `entityId`. Readers should use this directly — it is already
-    /// the name to show, so no per-read registry lookup is needed.
-    public let name: String
-    public let icon: String?
-    public let rawDeviceClass: String?
-
-    public init(
-        id: String,
-        entityId: String,
-        serverId: String,
-        domain: String,
-        name: String,
-        icon: String?,
-        rawDeviceClass: String?,
-    ) {
-        self.id = id
-        self.entityId = entityId
-        self.serverId = serverId
-        self.domain = domain
-        self.name = name
-        self.icon = icon
-        self.rawDeviceClass = rawDeviceClass
-    }
-
-    public var deviceClass: DeviceClass {
+// `HAAppEntity` itself lives in the `HAModels` package; these are its `DeviceClass` helper and
+// database-backed queries.
+public extension HAAppEntity {
+    var deviceClass: DeviceClass {
         DeviceClass(rawValue: rawDeviceClass ?? "") ?? .unknown
-    }
-
-    public enum ConfigInclude {
-        case all
-        case hidden
-        /// Kept for source compatibility. Disabled entities are no longer stored (the entity
-        /// registry is sourced from `list_for_display`, which omits them), so this has no effect.
-        case disabled
     }
 
     /// Fetches app entities based on configuration filters.
     /// - Parameter include: Filter options - use `.all` to include everything, or combine `.hidden` and `.disabled` to
     /// include specific types
     /// - Returns: Array of filtered entities
-    public static func config(include: [ConfigInclude] = []) throws -> [HAAppEntity] {
+    static func config(include: [ConfigInclude] = []) throws -> [HAAppEntity] {
         try Current.database().read({ db in
             // If .all is specified, return everything
             if include.contains(.all) {
@@ -102,7 +52,7 @@ public struct HAAppEntity: Codable, Identifiable, FetchableRecord, PersistableRe
         })
     }
 
-    public static func entity(id: String, serverId: String) -> HAAppEntity? {
+    static func entity(id: String, serverId: String) -> HAAppEntity? {
         do {
             return try Current.database().read { db in
                 try HAAppEntity
@@ -114,11 +64,5 @@ public struct HAAppEntity: Codable, Identifiable, FetchableRecord, PersistableRe
             Current.Log.error("Error fetching entity \(id) for server \(serverId): \(error)")
         }
         return nil
-    }
-}
-
-public enum ServerEntity {
-    public static func uniqueId(serverId: String, entityId: String) -> String {
-        "\(serverId)-\(entityId)"
     }
 }
