+++ Sources/Shared/Panels/AppPanel.swift
@@ -2,34 +2,9 @@ import Foundation
 import GRDB
 import HAKit
 
-public struct AppPanel: Codable, FetchableRecord, PersistableRecord {
-    public var id: String = UUID().uuidString
-    public var serverId: String = ""
-    public var icon: String? = nil
-    public var title: String = ""
-    public var path: String = ""
-    public var component: String = ""
-    public var showInSidebar: Bool = true
-
-    public init(
-        id: String = UUID().uuidString,
-        serverId: String,
-        icon: String? = nil,
-        title: String,
-        path: String,
-        component: String,
-        showInSidebar: Bool
-    ) {
-        self.id = id
-        self.serverId = serverId
-        self.icon = icon
-        self.title = title
-        self.path = path
-        self.component = component
-        self.showInSidebar = showInSidebar
-    }
-
-    public static func panels(serverId: String) throws -> [AppPanel]? {
+// `AppPanel` itself lives in the `HAModels` package; these are its database-backed queries.
+public extension AppPanel {
+    static func panels(serverId: String) throws -> [AppPanel]? {
         try Current.database().read({ db in
             try AppPanel
                 .filter(
@@ -39,7 +14,7 @@ public struct AppPanel: Codable, FetchableRecord, PersistableRecord {
         })
     }
 
-    public static func panelsPerServer() throws -> [Server: [AppPanel]] {
+    static func panelsPerServer() throws -> [Server: [AppPanel]] {
         var panelsPerServer: [Server: [AppPanel]] = [:]
         for server in Current.servers.all {
             do {
